correcccion recetario

+++ backend-app/schema/recipes/ingredients.py
@@ -11,8 +11,8 @@ from datetime import time, datetime
 class IngredientsPost (BaseModel):
     name:str
     purchasing_unit_measure_id:int
-    purchasing_price:int
-    number_units_purchasing:Optional[int] = None
+    purchasing_price:float
+    number_units_purchasing:Optional[float] = None
     purchasing_format_id:Optional[int] = None
     net_gross_weight:float
     shrinkage_persent:Optional[float] = None
@@ -21,7 +21,7 @@ class IngredientsPost (BaseModel):
 class IngredientsUpdate (BaseModel):
     name:str
     purchasing_unit_measure_id:int
-    purchasing_price:int
+    purchasing_price:float
     net_gross_weight:float
 
    
@@ -38,12 +38,12 @@ class RecipeDataIngredients (BaseModel):
     ingredient_id:int
     recipe_data_sheet_id:int
     unit_measure_id:int
-    quantity:int
-    quantity_before_shrinkage:int
+    quantity:float
+    quantity_before_shrinkage:float
 
 class RecipeDataIngredientsUpdate (IngredientsPost):
     unit_measure_id:int
-    quantity:int
+    quantity:float
